docs(viewer-docs): update viewport doc
